interpolation: support parsing the contents of an interpolation
The goal of this patch is to support parsing the contents of an
interpolation. This is done by emitting three types of interpolation
from the external scanner, rather than one:

INTERPOLATION_NOBRACE_VARIABLE: "$foo"
INTERPOLATION_BRACE_VARIABLE:   "${foo}" or "${foo.upcase()}" or "${foo[0]}"
INTERPOLATION_EXPRESSION:       "${2 + 2}"

For the first two types we additionally scan for the variable name with
scan_interpolation_nosigil_variable. If a variable is found we emit a
zero-width token and scan for the variable name in the grammar DSL.

We add this new variable type to the normal variable, but only scan for
one at the beginning of an interpolation.