interpolation: support parsing the contents of an interpolation
The goal of this patch is to support parsing the contents of an
interpolation. This is done by moving brace style expansion, ${}, from
the external scanner to the tree-sitter DSL. This works, but the
generated parse tree is not entirely accurate, for instance "${foo}" is
parsed as:

(double_quoted_string
  (interpolation
    (name)))))

Where it ideally should be:

(double_quoted_string
  (interpolation
    (variable
      (name))))))

But, I think the patch is still valuable as is.