heredoc: improve heredoc parsing
This patch re-works heredoc parsing, so as to treat the opening tag as a
string like object, and the heredoc body as a separate object on the
line following the opening tag. This allows for parsing heredoc uses,
such as:

notice(@(FIRST), @(SECOND))
    This is the text for the first heredoc
    | FIRST
    This is the text for the second
    | SECOND

Portions of the initial patch were taken from the ruby parser,
https://github.com/tree-sitter/tree-sitter-ruby, then reworked for
Puppet.